refactor: split reference tax dtos
